Refactor pagination into PaginationBarComponent
