+++ frontend/src/app/service/search/pagination-bar/pagination-bar.component.css
@@ -1,0 +1,33 @@
+.pagination-bar {
+    display: flex;
+    justify-content: center;
+    gap: 8px;
+    margin-top: 1rem;
+    list-style: none;
+    padding: 0;
+}
+
+.page-item {
+    cursor: pointer;
+    padding: 6px 12px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    user-select: none;
+    color: #333;
+}
+
+.page-item.disabled {
+    cursor: not-allowed;
+    opacity: 0.5;
+}
+
+.page-item.active {
+    background-color: #007bff;
+    color: white;
+    font-weight: bold;
+    border-color: #007bff;
+}
+
+.page-item:hover:not(.disabled):not(.active) {
+    background-color: #e9ecef;
+}
