+++ frontend/src/app/service/search/pagination-bar/pagination-bar.component.html
@@ -1,0 +1,31 @@
+@if (totalPages() > 1) {
+    <nav aria-label="Page navigation">
+        <ul class="pagination-bar">
+            <li
+                    class="page-item"
+                    [class.disabled]="currentPage() === 0"
+                    (click)="currentPage() > 0 && goToPreviousPageFn()"
+            >
+                Previous
+            </li>
+
+            <li
+                    *ngFor="let page of getPaginationPages()"
+                    class="page-item"
+                    [class.active]="page === currentPage()"
+                    [class.disabled]="page === '...'"
+                    (click)="page !== '...' && goToPageFn(page)"
+            >
+                {{ page === '...' ? '…' : displayPageNumber(page) }}
+            </li>
+
+            <li
+                    class="page-item"
+                    [class.disabled]="currentPage() + 1 >= totalPages()"
+                    (click)="currentPage() + 1 < totalPages() && goToNextPageFn()"
+            >
+                Next
+            </li>
+        </ul>
+    </nav>
+}
