+++ frontend/src/app/service/search/search-services.component.html
@@ -19,37 +19,14 @@
                     [services]="services">
             </app-service-list>
 
-            @if (totalPages() > 1) {
-                <nav aria-label="Page navigation">
-                    <ul class="pagination-bar">
-                        <li
-                                class="page-item"
-                                [class.disabled]="currentPage() === 0"
-                                (click)="currentPage() > 0 && goToPreviousPage()"
-                        >
-                            Previous
-                        </li>
+            <app-pagination-bar
+                    [currentPage]="currentPage"
+                    [totalPages]="totalPages"
+                    [goToPageFn]="goToPage.bind(this)"
+                    [goToPreviousPageFn]="goToPreviousPage.bind(this)"
+                    [goToNextPageFn]="goToNextPage.bind(this)"
+            ></app-pagination-bar>
 
-                        <li
-                                *ngFor="let page of getPaginationPages()"
-                                class="page-item"
-                                [class.active]="page === currentPage()"
-                                [class.disabled]="page === '...'"
-                                (click)="page !== '...' && goToPage(page)"
-                        >
-                            {{ page === '...' ? '…' : displayPageNumber(page) }}
-                        </li>
-
-                        <li
-                                class="page-item"
-                                [class.disabled]="currentPage() + 1 >= totalPages()"
-                                (click)="currentPage() + 1 < totalPages() && goToNextPage()"
-                        >
-                            Next
-                        </li>
-                    </ul>
-                </nav>
-            }
         </div>
     </div>
 }
